+++ src/main/java/ua/epam/spring/hometask/service/interf/EventCounterStatsService.java
@@ -6,15 +6,15 @@ import ua.epam.spring.hometask.domain.Event;
  * Created by Oleksii_Kovetskyi on 4/8/2018.
  */
 public interface EventCounterStatsService {
-    long getAccessByNameCount(Event event);
+    int getAccessByNameCount(Event event);
 
     void incrementAccessByNameCount(Event event);
 
-    long getPriceQueryCount(Event event);
+    int getPriceQueryCount(Event event);
 
     void incrementPriceQueryCount(Event event);
 
-    long getTicketsBookedTimesCount(Event event);
+    int getTicketsBookedTimesCount(Event event);
 
     void incrementTicketsBookedTimesCount(Event event);
 }
